+++ migrations/000001_create_table.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE metrics (
                          id SERIAL PRIMARY KEY,
                          name VARCHAR(255) NOT NULL,
                          type_metrics VARCHAR(255) NOT NULL,
-                         delta INTEGER,
+                         delta bigint,
                          value DOUBLE PRECISION
 );
